+++ common/src/main/java/ddraig/net/customraces/CustomRaces.java
@@ -26,6 +26,7 @@ public class CustomRaces {
         FirstJoinHandler.init();
         ddraig.net.customraces.event.WereRaceTransformHandler.init();
         ddraig.net.customraces.event.RaceSoundHandler.init();
+        ddraig.net.customraces.event.MobAllianceHandler.init();
 
         // Passive ability player tick loop
         TickEvent.PLAYER_POST.register(player -> {